+++ packages/vscode/src/topicCreation.ts
@@ -1,0 +1,13 @@
+import * as vscode from 'vscode';
+
+export function buildTopicConfig() {
+	// query the user for the topic parameters
+	// return a topic config object
+	console.log('buildTopicConfig');
+}
+
+export function createTopicFolder(parentFolderUri: vscode.Uri) {
+	// create a new topic folder that includes an images directory etc.
+	// add the topic folder to the list of topic folders that can be built
+	console.log('createTopicFolder');
+}
